refactor: 패키지 변경 및 코드 리펙토링
