+++ src/main/java/kr/co/antoon/oauth/presentation/AuthController.java
@@ -12,7 +12,8 @@ import javax.servlet.http.HttpServletResponse;
 @Slf4j
 @RestController
 @RequiredArgsConstructor
-public class OauthController {
+@RequestMapping("/api/v1/auth")
+public class AuthController {
     private final AuthService authService;
 
     @GetMapping("/refresh")
