+++ src/main/java/kr/co/antoon/user/domain/vo/Role.java
@@ -7,7 +7,6 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public enum Role {
 
-    GUEST("ROLE_GUEST", "손님"),
     USER("ROLE_USER", "일반 사용자");
 
     private final String key;
